Allow `m_sensitivity` CVAR to be a float
More testing required, but seems to work well. Will need to strip all trailing zeros in doomretro.cfg. Same change need to be applied to game controller's sensitivity CVARs, and maybe others. Thanks to Endless for the suggestion. https://www.doomworld.com/forum/post/2561342

+++ src/g_game.c
@@ -851,8 +851,8 @@ bool G_Responder(event_t *ev)
 
             if (!automapactive || am_followmode)
             {
-                mousex = ev->data2 * m_sensitivity / 10;
-                mousey = -ev->data3 * m_sensitivity / 10;
+                mousex = (int)(ev->data2 * m_sensitivity / 10.0f);
+                mousey = (int)(-ev->data3 * m_sensitivity / 10.0f);
             }
 
             return true;    // eat events
